Moved dragging logic to new file
