+++ client/src/components/main-content/SectionList.jsx
@@ -1,7 +1,6 @@
 import { useState, useContext } from "react";
 import { DndContext, DragOverlay } from "@dnd-kit/core";
 import {
-  arrayMove,
   SortableContext,
   horizontalListSortingStrategy,
   verticalListSortingStrategy,
@@ -15,13 +14,14 @@ import DroppableSection from "./Sections/DroppableSection";
 import ItemModal from "./Items/ItemModal";
 import SectionModal from "./Sections/SectionModal";
 import "./SectionList.css";
-import {
-  DraggableComponentTypes,
-  SectionActions,
-  DragEndActions,
-} from "../../utils/constants";
+import { SectionActions } from "../../utils/constants";
 import customCollisionDetection from "../../utils/customCollisionDetection";
 import ViewContext from "../../ViewContext";
+
+import {
+  handleDragEnd as handleDragEndUtil,
+  findItemBySection,
+} from "../../utils/sectionListUtils";
 
 const SectionList = () => {
   const { viewMode } = useContext(ViewContext);
@@ -61,165 +61,21 @@ const SectionList = () => {
   const [targetSectionId, setTargetSectionId] = useState(null);
   const [editingItem, setEditingItem] = useState(null);
 
-  const findItemBySection = (section) => {
-    for (const i of section.items) {
-      if (i.id === activeId) {
-        return i;
-      }
-    }
-
-    return null;
-  };
-
   const handleDragStart = (event) => {
     setActiveId(event.active.id);
   };
 
-  const handleDragEndSection = (active, over) => {
-    if (active.id !== over.id) {
-      setSectionOrder((prev) => {
-        const oldIndex = prev.indexOf(active.id);
-        const newIndex = prev.indexOf(over.id);
-        return arrayMove(prev, oldIndex, newIndex);
-      });
-    }
-    setActiveId(null);
-  };
-
-  const handleDragEndItem = (active, over) => {
-    const fromSectionId = active.data.current.sectionId;
-    const toSectionId = over.data.current?.sectionId || over.id;
-    if (!toSectionId) {
-      setActiveId(null);
-      return;
-    }
-
-    if (fromSectionId === toSectionId && active.id === over.id) {
-      setActiveId(null);
-      return;
-    }
-
-    const item = findItemBySection(sections[fromSectionId]);
-
-    setSections((prev) => {
-      // remove from old section
-      const newFromItems = prev[fromSectionId].items.filter(
-        (i) => i.id !== active.id
-      );
-
-      const filteredToItems = prev[toSectionId].items.filter(
-        (i) => i.id !== active.id
-      );
-
-      const overIndex = filteredToItems.findIndex((i) => i.id === over.id);
-      let newToItems;
-      if (overIndex === -1) {
-        newToItems = [...filteredToItems, item];
-      } else {
-        newToItems = [
-          ...filteredToItems.slice(0, overIndex),
-          item,
-          ...filteredToItems.slice(overIndex),
-        ];
-      }
-
-      return {
-        ...prev,
-        [fromSectionId]: { ...prev[fromSectionId], items: newFromItems },
-        [toSectionId]: { ...prev[toSectionId], items: newToItems },
-      };
-    });
-    setActiveId(null);
-  };
-
-  const handleDragEndAdd = (active, over) => {
-    setShowModal(true);
-    setActiveId(null);
-  };
-
-  const handleDragEndDelete = (active, over) => {
-    // Delete item
-    if (active.data.current?.type === DraggableComponentTypes.ITEM) {
-      const fromSectionId = active.data.current.sectionId;
-      setSections((prev) => ({
-        ...prev,
-        [fromSectionId]: {
-          ...prev[fromSectionId],
-          items: prev[fromSectionId].items.filter((i) => i.id !== active.id),
-        },
-      }));
-    }
-    // Delete section
-    if (active.data.current?.type === DraggableComponentTypes.SECTION) {
-      setSections((prev) => {
-        const newSections = { ...prev };
-        delete newSections[active.id];
-        return newSections;
-      });
-      setSectionOrder((prev) => prev.filter((id) => id !== active.id));
-    }
-    setActiveId(null);
-    return;
-  };
-
   const handleDragEnd = (event) => {
-    const { active, over } = event;
-    if (!over) {
-      setActiveId(null);
-      return;
-    }
-
-    let actionType = null;
-
-    if (over.id === SectionActions.DELETE_ZONE) {
-      actionType = DragEndActions.DELETE;
-    } else if (
-      over.id === SectionActions.DROPZONE &&
-      active.id === SectionActions.ADD
-    ) {
-      actionType = DragEndActions.ADD_SECTION;
-    } else if (
-      active.id === SectionActions.ADD &&
-      (over.data?.current?.type === DraggableComponentTypes.SECTION ||
-        over.data?.current?.type === DraggableComponentTypes.ITEM)
-    ) {
-      actionType = DragEndActions.ADD_ITEM;
-    } else if (
-      active.data.current?.type === DraggableComponentTypes.SECTION &&
-      over.data.current?.type === DraggableComponentTypes.SECTION
-    ) {
-      actionType = DragEndActions.MOVE_SECTION;
-    } else if (active.data.current?.type === DraggableComponentTypes.ITEM) {
-      actionType = DragEndActions.MOVE_ITEM;
-    }
-
-    switch (actionType) {
-      case DragEndActions.DELETE:
-        handleDragEndDelete(active, over);
-        break;
-      case DragEndActions.ADD_SECTION:
-        handleDragEndAdd(active, over);
-        break;
-      case DragEndActions.ADD_ITEM:
-        {
-          const sectionId =
-            over.data.current.type === DraggableComponentTypes.SECTION
-              ? over.id
-              : over.data.current.sectionId;
-          setTargetSectionId(sectionId);
-          setShowItemModal(true);
-          setActiveId(null);
-        }
-        break;
-      case DragEndActions.MOVE_SECTION:
-        handleDragEndSection(active, over);
-        break;
-      case DragEndActions.MOVE_ITEM:
-        handleDragEndItem(active, over);
-        break;
-      default:
-        setActiveId(null);
-    }
+    handleDragEndUtil(event, {
+      setActiveId,
+      activeId,
+      setShowModal,
+      setShowItemModal,
+      setTargetSectionId,
+      setSections,
+      setSectionOrder,
+      sections,
+    });
   };
 
   const handleSaveSection = () => {
@@ -303,7 +159,7 @@ const SectionList = () => {
     }
     // for items
     for (const section of Object.values(sections)) {
-      let item = findItemBySection(section);
+      let item = findItemBySection(section, { activeId });
 
       if (item) {
         return (
@@ -360,10 +216,7 @@ const SectionList = () => {
             collisionDetection={customCollisionDetection}
             onDragOver={handleDragOver}
           >
-            <SortableContext
-              items={sectionOrder}
-              strategy={sortStrategy}
-            >
+            <SortableContext items={sectionOrder} strategy={sortStrategy}>
               <div
                 className={`sections-row ${
                   viewMode === "list" ? "list-view" : "board-view"
