+++ client/src/utils/sectionListUtils.js
@@ -1,0 +1,194 @@
+import {
+  SectionActions,
+  DragEndActions,
+  DraggableComponentTypes,
+} from "./constants";
+import { arrayMove } from "@dnd-kit/sortable";
+
+export const findItemBySection = (section, { activeId }) => {
+  for (const i of section.items) {
+    if (i.id === activeId) {
+      return i;
+    }
+  }
+  return null;
+};
+
+const handleDragEndSection = (
+  active,
+  over,
+  { setSectionOrder, setActiveId }
+) => {
+  if (active.id !== over.id) {
+    setSectionOrder((prev) => {
+      const oldIndex = prev.indexOf(active.id);
+      const newIndex = prev.indexOf(over.id);
+      return arrayMove(prev, oldIndex, newIndex);
+    });
+  }
+  setActiveId(null);
+};
+
+const handleDragEndItem = (
+  active,
+  over,
+  { setSections, setActiveId, sections, activeId }
+) => {
+  const fromSectionId = active.data.current.sectionId;
+  const toSectionId = over.data.current?.sectionId || over.id;
+  if (!toSectionId) {
+    setActiveId(null);
+    return;
+  }
+
+  if (fromSectionId === toSectionId && active.id === over.id) {
+    setActiveId(null);
+    return;
+  }
+
+  const item = findItemBySection(sections[fromSectionId], { activeId });
+
+  setSections((prev) => {
+    // remove from old section
+    const newFromItems = prev[fromSectionId].items.filter(
+      (i) => i.id !== active.id
+    );
+
+    const filteredToItems = prev[toSectionId].items.filter(
+      (i) => i.id !== active.id
+    );
+
+    const overIndex = filteredToItems.findIndex((i) => i.id === over.id);
+    let newToItems;
+    if (overIndex === -1) {
+      newToItems = [...filteredToItems, item];
+    } else {
+      newToItems = [
+        ...filteredToItems.slice(0, overIndex),
+        item,
+        ...filteredToItems.slice(overIndex),
+      ];
+    }
+
+    return {
+      ...prev,
+      [fromSectionId]: { ...prev[fromSectionId], items: newFromItems },
+      [toSectionId]: { ...prev[toSectionId], items: newToItems },
+    };
+  });
+  setActiveId(null);
+};
+
+const handleDragEndAdd = (active, over, { setShowModal, setActiveId }) => {
+  setShowModal(true);
+  setActiveId(null);
+};
+
+const handleDragEndDelete = (
+  active,
+  { setSections, setSectionOrder, setActiveId }
+) => {
+  // Delete item
+  if (active.data.current?.type === DraggableComponentTypes.ITEM) {
+    const fromSectionId = active.data.current.sectionId;
+    setSections((prev) => ({
+      ...prev,
+      [fromSectionId]: {
+        ...prev[fromSectionId],
+        items: prev[fromSectionId].items.filter((i) => i.id !== active.id),
+      },
+    }));
+  }
+  // Delete section
+  if (active.data.current?.type === DraggableComponentTypes.SECTION) {
+    setSections((prev) => {
+      const newSections = { ...prev };
+      delete newSections[active.id];
+      return newSections;
+    });
+    setSectionOrder((prev) => prev.filter((id) => id !== active.id));
+  }
+  setActiveId(null);
+  return;
+};
+
+export const handleDragEnd = (
+  event,
+  {
+    setActiveId,
+    activeId,
+    setShowModal,
+    setShowItemModal,
+    setTargetSectionId,
+    setSections,
+    setSectionOrder,
+    sections,
+  }
+) => {
+  const { active, over } = event;
+  if (!over) {
+    setActiveId(null);
+    return;
+  }
+
+  let actionType = null;
+
+  if (over.id === SectionActions.DELETE_ZONE) {
+    actionType = DragEndActions.DELETE;
+  } else if (
+    over.id === SectionActions.DROPZONE &&
+    active.id === SectionActions.ADD
+  ) {
+    actionType = DragEndActions.ADD_SECTION;
+  } else if (
+    active.id === SectionActions.ADD &&
+    (over.data?.current?.type === DraggableComponentTypes.SECTION ||
+      over.data?.current?.type === DraggableComponentTypes.ITEM)
+  ) {
+    actionType = DragEndActions.ADD_ITEM;
+  } else if (
+    active.data.current?.type === DraggableComponentTypes.SECTION &&
+    over.data.current?.type === DraggableComponentTypes.SECTION
+  ) {
+    actionType = DragEndActions.MOVE_SECTION;
+  } else if (active.data.current?.type === DraggableComponentTypes.ITEM) {
+    actionType = DragEndActions.MOVE_ITEM;
+  }
+
+  switch (actionType) {
+    case DragEndActions.DELETE:
+      handleDragEndDelete(active, over, {
+        setSections,
+        setSectionOrder,
+        setActiveId,
+      });
+      break;
+    case DragEndActions.ADD_SECTION:
+      handleDragEndAdd(active, over, { setShowModal, setActiveId });
+      break;
+    case DragEndActions.ADD_ITEM:
+      {
+        const sectionId =
+          over.data.current.type === DraggableComponentTypes.SECTION
+            ? over.id
+            : over.data.current.sectionId;
+        setTargetSectionId(sectionId);
+        setShowItemModal(true);
+        setActiveId(null);
+      }
+      break;
+    case DragEndActions.MOVE_SECTION:
+      handleDragEndSection(active, over, { setSectionOrder, setActiveId });
+      break;
+    case DragEndActions.MOVE_ITEM:
+      handleDragEndItem(active, over, {
+        setSections,
+        setActiveId,
+        sections,
+        activeId,
+      });
+      break;
+    default:
+      setActiveId(null);
+  }
+};
